refactoring bulk reschedule in enrolment

+++ backend/controllers/LessonController.php
@@ -567,7 +567,9 @@ class LessonController extends BaseController
             foreach ($lessons as $i => $lesson) {
                 $oldLesson = Lesson::findOne($oldLessonIds[$i]);
                 $oldLesson->rescheduleTo($lesson);
-                LessonHierarchy::deleteAll($oldLesson->id);
+                if ($oldLesson) {
+                    LessonHierarchy::deleteAll($oldLesson->id);
+                }
                 $bulkReschedule = new BulkRescheduleLesson();
                 $bulkReschedule->lessonId = $lesson->id;
                 $bulkReschedule->save();
